+++ docs/about/release-notes.rst
@@ -41,10 +41,12 @@ v0.xy.0 (Unreleased)
 Features
 ~~~~~~~~
 
+* Add :py:func:`scipp.elemwise_func` for defining custom transformations of input variables based on a Python function compiled via ``numba.cfunc`` `#2886 <https://github.com/scipp/scipp/pull/2886>`_.
+
 Breaking changes
 ~~~~~~~~~~~~~~~~
 
-* The SciPy wrappers ``integrate``, ``interpolate``, ``ndimage``, ``optimize``, and ``signal`` were moved into the :py:module:`scipp.scipy` submodule `#2881 <https://github.com/scipp/scipp/pull/2881>`_.
+* The SciPy wrappers ``integrate``, ``interpolate``, ``ndimage``, ``optimize``, and ``signal`` were moved into the :py:mod:`scipp.scipy` submodule `#2881 <https://github.com/scipp/scipp/pull/2881>`_.
 
 Bugfixes
 ~~~~~~~~